msm: subsystem_restart: Send sysmon event notifications on powerup
Currently, when a subsystem powers up, there is no mechanism
to let it know about the state of the other online subsystems.

Notifications regarding the current status of other subsystems
are sent to the subsystem being brought out of reset/restarted.
This is done immediately after powering up the subsystem. What
this subsystem does with the notification info is left upto it.

Change-Id: I8d9794babd26dcec7af0ee868f66c7fb4c5c8a34
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>