char: msm_smd_pkt: Add warning to driver removal function
The current implementation of smd_pkt_remove_driver does
not check to see if it has found the driver to be removed
in the module's list of drivers before attempting removal.

Add a flag that indicates whether or not the driver
was found in the list before removal is attempted. If
no driver is found, an error message is logged, and the
removal is not attempted.

Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>