soc: qcom: smp2p: Use memcpy_toio() to copy to shared memory
Shared memory is device type memory which has strict alignment
requirements.  Standard *cpy functions like memcpy() and strcpy() do not
guarantee the correct alignment.  The correct *cpy functions to access
device memory with are memcpy_toio() and memcpy_fromio().

Replace a strlcpy() call that was copying a string to shared memory with
a call to memcpy_toio() to guarentee the accesses to device memory are
correctly aligned.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>