ARM: Speed up page list initialization during boot
During boot, we populate the page lists by using the page freeing
mechanism on every individual page.  Unfortunately, this is very
inefficient because the memory manager spends a lot of time coalescing
pairs of adjacent free pages into bigger blocks.

Rather than adding a single page at a time, we can take advantage of the
fact that we know that all the pages are available and free up big
blocks of pages at a time instead.

BUG=chrome-os-partner:26945
TEST=platform_BootPerf, seconds_power_on_to_kernel

Change-Id: Ifaad5df1bc310e5e675a332cff8cdad5b3011055
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/190381
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>