msm: tspp: Remove tspp driver user-space API
tspp driver user-space API is not in fact used and
therefore was not maintained so it is obsolete.
With 64bit porting in mind it was decided this driver
will no longer be visible as char device to user.
The driver will be operated via the dvb/demux device
and user-space API.

Change-Id: Ib6a9240a211334bd968a3558bd1783a24aeacda7
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>