msm: kgsl: Add IFH feature by removing IB commands
Removing the IBs from being sent to the ringbuffer limits the
actual amount of work that the GPU executes. This allows for
a more realistic way to show where the CPU bottlenecks are. By
removing just the IBs and not modifiying anything else, the GPU
still executes some basic commands.  More importantly, it will
still execute the interrupt path which will give more realistic
representations of where the CPU actually spends time.

Signed-off-by: Carter Cooper <ccooper@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>