ARM: dts: msm: Reserve early bootloader memory on 8084
On most systems, the early bootloader memory overlaps with regions
that are already removed for other purposes. The memory map on 8084
does not have the early bootloader memory already reserved. Because
the early bootloaders need to run to collect ramdumps, if this
memory is not reserved, ram dumps may get corrupted. Make sure this
region of memory is reserved.

Signed-off-by: Laura Abbott <lauraa@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>