esoc: mdm-4x: Add reboot notification
Add a reboot notifier that will enumerate through the list of registered
modems and try to perform a safe shutdown on them, before finally
driving the AP2MDM_SOFT_RESET line to shut off the modem. This is needed
to ensure the modem actually powers off when the AP does. This will only
handle the graceful shutdowns, any abnormal crash will need a
corresponding change in the bootloaders to shut them off upon first
boot.

Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: Christopher Fries <c.fries@motorola.com>