msm: kgsl: Improve the CFF memory dump functions
kgsl_cffdump_setmem really wanted to be a memset clone but it
was reading in what was essentially a 8 bit char and writing
a 32 bit dword with the end result being that the data written
in the CFF wasn't matching the data written to memory. On top of
this the CFF code really went out of its way to try
handle unaligned sizes but it only prints memory data in dword
increments. Remove the extra code and just print the incoming
memory in dword groups as nature intends.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>