msm: mdss: remove wait for pingpong during status check
The wait for pingpong is only necessary if using the method of triggering
BTA to check the panel status.  Since we don't use this method, the wait
for pingpong can be removed.

Since we don't need to wait for pingpong, it is ok to remove the lock of
mdp5_data->ov_lock and allow new frames to be submitted during the status
check.  All locking necessary for sending commands for the status check
while continuing to send new frames is already handled.

Signed-off-by: Patrick Auchter <auchter@motorola.com>
Reviewed-by: Fred Fettinger <fettinge@motorola.com>
Reviewed-by: Jared Suttles <jsuttles@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>