uio: msm_sharedmem: Add support for dynamic shared memory allocation
The shared memory regions used for the UIO devices use physical
addresses that are hardcoded in the device tree which creates a
dependency on the memory map.

Add support to dynamically allocate uncached physically contiguous
memory for the shared buffer at device probe and allow clients (remote-
fs and rfsa) to request this shared buffer address through Kernel QMI
and remove this dependency.

Change-Id: I481f0863db02964336fd461045b5e944ae4f8f2a
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>