qcom: subsys-pil-tz: Add qcom prefix to subsystem dt properties
Some of the properties used by the subsys-pil-tz driver are
qcom specific. Add "qcom" prefix to these.

CRs-Fixed: 662297
Change-Id: I8c6fa8daa18f5337d64064eec116d81f185d08bb
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>

msm: vidc: Add driver to bring Venus subsystem out of reset

Add a driver that is capable of bringing the Venus subsystem out
of reset when the "qcom,pil-no-auth" boolean is set in the Venus
device node in DT. This driver depends on the PIL driver to load
the Venus firmware image into memory.

Change-Id: I46667ac24f8769ea1366c3c04447c912916cc6ca
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Shalaj Jain <shalajj@codeaurora.org>
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>