net: ipc_router: Reference count local port
Everytime a port is searched and then accessed, a global lock is
used to protect the port for the entire duration of the access.

Define a reference element in the port. Get a reference to the port
everytime it is searched and accessed. Put the reference back once the
access is complete. Release the port when the reference count is zero.

Update the lock hierarchies so that the changes do not cause out of order
locking.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>