msm: mdss: Use ULPS mode for partial mode
Seeing issues with higher than expected current drain when kernel hands off
control of the display for partial mode.  This is due to not all the MIPI
lanes being in a state that the display expects.
To address this, after we hand control of the switched MIPI lanes, enter
ULPS mode on the host side.  This will cause the unswitched MIPI lanes
to go to a state which allows the display to go to a lower power state.

Signed-off-by: Bang Nguyen <bnguyen1@motorola.com>

Conflicts:
	drivers/video/msm/mdss/mdss_dsi.c
	drivers/video/msm/mdss/mdss_dsi.h
	drivers/video/msm/mdss/msm_mdss_io_8974.c