ASoC: compress: fix the target not going to TCXO shutdown
The error code propagated from the flush cmd in the trigger stop
is causing the error in the state of the soc compress framework.
This hinders in closing the BE connected to the FE.
If the AFE is not closed it prevents the target to go into TCXO
shutdown.
Fixed by removing the error handling of the flush command in the
trigger stop, since failure in flush doesn't signify the
failure in the stop command.

Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>
[neo: Also cleanup mismerge of 38d8ba0cc229b26f3e11340710af09179655d8ca]
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>