qcacld: From HDD remove the logic to try shared auth if open fails.
HDD tries to connect using shared auth if connect with open auth
fails. This leads driver to send 2 Assoc reject to the
supplicant which may lead to mismatch in states of driver and
supplicant.

In LIM if auth type is eCSR_AUTH_TYPE_AUTOSWITCH and if the open
auth is rejected with reason code
eSIR_MAC_AUTH_ALGO_NOT_SUPPORTED_STATUS (13) it try again with
shared auth.
Removes the logic from HDD as it is already handled by LIM.

Change-Id: If0dae1d3aadf3c833b7ce3064b04febfbeedc233
CRs-Fixed: 689892