lib: qmi_encdec: Add support for string data type
Add support to encode and decode string data type into a QMI message.
If a string has a <Type>:<Length>:<Value> information associated with it,
then the length information holds the length of the string. If the string
is part of a nested structure, then the length of the string is explicitly
encoded into the QMI message.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
[neo: Patch modified to not define U8_MAX again.]
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>