msm-core: Supplement scheduler core selection using power estimates
The msm-core driver provides per-core temperature and power for every
P-state (frequency/capacity) to the scheduler.
The objective of the driver is to compute the power consumed
at every P-state so as to make power efficient scheduling decisions.
     Dynamic power = fn (V, f)
     where V is voltage at the frequency f

The power computation and temperature read for all cores happen at
regular intervals (deferrable) unless there is a change in temperature
or frequency.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>