qcom: subsystem_restart: Skip collecting ramdumps if a gpio bit is set
A feature in the modem requires that the modem be
restarted as part of normal operation rather than
as a response to a fatal error. In these instances,
it is preferred that ramdump collection be avoided.
If a certain GPIO is set, disable ramdump collection
during a subsystem restart sequence.

Change-Id: I5eaedf8f43c06124cc30c3e8f5e8ed882e25dd49
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>