msm-core: Update frequency table when new policy is created
If the number of boot cpus is lesser than the total number of possible
cpus, the driver fails as all the cpus do not have frequency table.

Get the frequency table for online cpus during probe and update the
table for the rest when the policy is created for these cpus.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>