msm: mdss: hdmi: continuous splash screen for HDMI
Enable continuous splash screen for HDMI interface
when it is configured as primary. Enable the regulators, clocks
and gpios at probe to vote for them. If resolution change has
happened, power off the HDMI to clean the obsolete resolution
data and setting. If resolution has not changed, do not re-program
same settings for smoother display.

Change-Id: I45e76ffce8f0654673a8fb19b683c783a18e229b
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>