cnss: Dynamically allocate ramdump memory
Existing solution to reserve the memory region for CNSS ramdump
collection has it's own limitations because of reserving the
memory region for each platforms even before CNSS driver is
probed. Add implementation to allocate the required memory for
CNSS ramdump using dma_alloc_coherent API and update the physical
address in dump table so that tools can find the starting address
of the CNSS ramdump in the crash dump.

Change-Id: I66a5abe098852115a11b0b4231d3692cf79da75c
Signed-off-by: Prashanth Bhatta <bhattap@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>