iommu: msm: actually disable L2 redirect when CONFIG_IOMMU_PGTABLES_L2=n
We recently moved the L2 redirect configuration to a domain attribute
[dffd6f05e94f: "iommu: msm: move L2 redirect to a domain attribute"].
This allows us to do away with some divergence in the generic IOMMU
APIs.  However, that commit botched the case when
CONFIG_IOMMU_PGTABLES_L2 is disabled.  In that case we actually always
set the page tables to be shared, exactly the opposite of what we want.
Fix this.

Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>