ASoC: wcd9xxx: Handle codec register config if MBHC is disabled
On codec reset, mic and micbias pulldown bits are enabled
by default. In cases when MBHC block on the codec is not
enabled, the pulldown bits should be disabled so that mic
and micbias can be enabled for voice call use cases on Headset.
Add support to check whether MBHC is enabled and update the
pulldown bits accordingly. Also, it is possible that DCE_Z
and DCE_MB calibration values can be zero if MBHC block is not
enabled, add check to avoid division by zero in this scenario.

Signed-off-by: Phani Kumar Uppalapati <phaniu@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>