mfd: wcd9xxx-slimslave: Return error to caller if buffer timeout happens
For the apps-owned data port (master port), in order to receive the data
on the channel, buffers are queued to slimbus/BAM pipe. Under unexpected
error scenarios, like SSR for example, the buffer is never returned,
which causes timeout on the wait for buffer completion. Fix to add error
return under such case so that the caller can take necessary action and
not queue any more buffers as the underlying driver/hardware is in
erroneous state.

Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>