soc: qcom: peripheral-loader: Don't use virt addresses in do_elf_ramdump
Relocatable PIL ensures that there is no kernel mapping
present for PIL images, in order to prevent accesses to
those regions whilst they are owned by the subsystem.
Therefore don't attempt to pass in a non-existent mapping
to the ramdump API.

Also ensure that the segment structures are allocated
with kcalloc instead of kmalloc_array in order to ensure
zeroing out of other fields in the segment structures.

A future improvement would be to pass in the device pointer
used to do the dma_remap by peripheral-loader (while actually
loading the image) to the ramdump API to actually allow the
latter to do the remap on its own.

Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
Reviewed-on: http://gerrit.mot.com/690540
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Matthew Deheck <w15756@motorola.com>
Reviewed-by: Ling Jin <lingjin@motorola.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>