cnss: Enable SSR even when no ramdump memory
When there is no ramdump memory available, Subsystem Restart(SSR)
should work as expected. But if ramdump memory is not available
then subsystem device node would be offline because of not
calling subsystem_get. This would cause SSR to be aborted when
CNSS crash is detected. Fix the issue by calling subsystem_get
even when ramdump memory is not available where in CNSS SSR would
be triggered just that ramdump wouldn't be collected.

CRs-fixed: 762694
Change-Id: I65d867018ac99adca5c27056660dbefde34d93ca
Signed-off-by: Prashanth Bhatta <bhattap@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>