msm-core: Set up thresholds only until the hotplug temperature point
Currently, we set up thermal thresholds to notify of every 5C rise in
temperature in the positive direction, to help recalculate energy
values. When temperatures ramp quickly, setting up additional thresholds
for EA, seems to interfere with the ability of the thermal sensor
framework to quickly notify the thermal mitigation handlers of the high
temperature, preventing hotplug of the thermally sensitive core.

Fix this by allowing thersholds to be set only until the operating range
of the cpu.

Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>