qcom: ssr: Add the subsystem_get_with_fw() API
On newer targets having a common single binary, subsystem
clients will need the option to specify the firmware image
to use for booting up the subsystem. Provide a new API for
the same.
Clients will now be able to specify the firmware name to use
as one of the parameters to this API. Also, for clients which
trigger subsystem boot and shutdown via the open and close file
ops, a sysfs node "firmware_name" is exposed. By writing to this
node, the client can change the firmware used to boot dynamically.

Change-Id: I294bd8c23bb208d0f7168e7bd7d81980223275d2
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>