hwspinlock/msm: Add PID tracking for remote spinlocks
The msm_remote_spinlock driver currently does not store the owner of the
lock (the raw value of the lock), and the state of the hardware lock is not
saved in memory dumps. This increases difficulty when debugging situations
where the lock could not be acquired.

Store the value of each lock in an array, and provide an API to read
elements from this array for testing purposes.

Change-Id: I7186c5201bbd6e51249fb073e899d2ac69e5bec7
Signed-off-by: Steven Cahail <scahail@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>