msm: thermal: Limit the emergency mitigation messages
Thermal can set tsens threshold even if the threshold
value is equal to the current temperature. This will
trigger a tsens notification immediately, which will
print a kernel message. If the temperature stays in
the current value then the emergency mitigation
messages will flood the kernel logs.

Use rate limit feature in printk to reduce the amount
of messages printed from thermal.

Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

 Conflicts:
	drivers/thermal/msm_thermal.c