SPI: spi_qsd: Fix unbalanced clocks
While using runtime pm if there is an error getting
local resources the error code wasn't being
propagated up to the runtime pm framework causing
the runtime pm framework to think the device was
active when it wasn't.  This would cause runtime pm
to call the suspend callback causing unbalanced clocks.
This patch propagates the error back to the runtime
pm framework so the runtime pm state mirrors the actual
state of the hardware.

Signed-off-by: Dan Sneddon <dsneddon@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>