esoc: mdm-4x: Add support to send shutdown request
The shutdown request must be sent from mdm driver
to prevent a race condition between the request being
processed by the mdm followed by its reset and the
APQ disabling the monitoring the mdms status line.
If the request was sent by SSR, in some cases, the mdm
would shutdown, before APQ could disable the status
interrupt, and it would be interpreted by APQ as
an erroneous hard reset of the md.

Change-Id: I413df861ff5ef413a84fdf4778829e8af0cf8df5
Signed-off-by: Hanumant Singh <hanumant@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>