qcacld: Abort scan for SSID if deauth is received from supplicant.
During connection if AP is not found in scan list, driver tries
to scan for the ssid (eCsrScanForSsid) and if AP is found,
the connect cmd is started.

During this scan if Disconnect comes from supplicant, as there is
no ongoing connect command and sme is not in connected state, the
disconnect command is dropped in SME and hdd is indicated success.
This leads to HDD waiting on the event disconnect_comp_var and
connState to eConnectionState_Disconnecting.
After this if supplicant send connect and as HDD is in
disconnecting state it will again wait for disconnect_comp_var
and return failure after timeout and this will continue and user
will not able to connect.

If disconnect is received during eCsrScanForSsid abort/dequeue
the eCsrScanForSsid in SME and in HDD set the connState to
eConnectionState_NotConnected and set disconnect_comp_var
complete.

Change-Id: I1da89065e1e59bcf08806179f85e3d9d00d5828e
CRs-Fixed: 689890