ASoC: msm: qdsp6v2: check audio client pointer before accessing
In the registered callback for q6asm, we are checking if
the audio client pointer is valid and also, dereferencing it
to get the session ID even though it could be invalid or expired.
Return and exit immediately if the audio client pointer is
invalid.

Signed-off-by: Aravind Kumar <akumark@codeaurora.org>
Reviewed-on: http://gerrit.mot.com/764819
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Qiang Qiu <w20252@motorola.com>
Reviewed-by: Hui Dai <briandai@motorola.com>
Reviewed-by: Vladimir Karpovich <vkarpovich@motorola.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>