msm: mdss: rotator: ignore panel status for rotator fences
In case of rotator buf sync ioctl, the ioctl could be called on
frame buffer device for which panel is off. However rotator sessions
may still be active. In such cases the call shouldn't fail, thus move
check for panel off only when the target sync fence timeline is for
the panel.

Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>