sched: optimize select_best_cpu() with same power cost CPU mask
At present, the function select_best_cpu() has to continue to iterate the
search CPUs for the possibility of finding a more power efficient CPU even
if the best CPU based on cstate already found.  We can avoid this
unnecessary search if it is known ahead of time that CPUs in a cluster
have the same power efficiency.

Introduce kernel parameter 'sched_same_power_cost_cpus' to inform the
CPU mask that has an identical power costs and optimize the 'for' loop of
select_best_cpu() by ruling out CPUs that cannot be more power efficient
than cstate based best CPU with the same power cost CPU mask.

Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>