msm: ADSPRPC: Enable remote heap with SSR support
The remote heap feature allows for the remote processor to
request memory from application subsystem for use cases where
it does not have enough memory available locally and later
release the memory when use case ends. This memory is XPU
protected on allocation and in cases of subsystem restart this
memory would need to be unlocked and freed.

Acked-by: Himateja Reddy <hmreddy@qti.qualcomm.com>
Signed-off-by: Sathish Ambley <sathishambley@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>