net: Copy ndisc_nodetype from original skb in skb_clone
skb_clone does not copy all fields from original skb. Fields
which are uninitialized may cause unexpected behavior when they
are passed down to network stack. In case ndisc_nodetype is set
to a random value for a router advertisement, kernel looks at this
field and drops it assuming that this packet originates from a
node and not a router.

Fix this by copying the ndisc_nodetype from the original skb
rather than relying on an uninitialized cloned skb.

Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>