soc: qcom: qmi: Fix race during service event notification
Currently the service event notification is done based on the number of
services of <service:instance> type available. If a service of a specific
type exits and registers in quick succession, then the service event
notification gets suppressed without client's knowledge. The client then
tries to communicate with a stale/non-existent service.

Fix the race condition by maintaining a list of port addresses of all
available services.

Change-Id: I7f626c4bb2d3429b1b277d802e23e247100c371a
Signed-off-by: Atish Kumar Patra <apatra@codeaurora.org>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>