msm: vidc: Allow clients to specify/query the colorspace
When encoder clients specify the color space, the resulting bitstream
will have the color space embedded in it.  This allows for the
decoder to display the picture accurately without the color being
skewed.

Similarly decoder clients can query the color space of the bitstream by
enabling MSM_VIDC_EXTRADATA_VUI_DISPLAY_INFO.  If the bitstream contains
the colorspace hint, the appropriate extradata will be emitted by the
decoder.

Author: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Praveen Chavan <pchavan@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>